Fix stylesheet link tag
The all.css.sass stylesheet imports the other stylesheets and only that
file should be linked from the layout.

Currently each individual stylesheet is compiled to sass and copied over
to the build directory. This is no harm - but not the desired
functionality either. Will have a look at that later.